Reorg calendar into domains/ (maximal split)
The most entangled domain — shared event validators, the calendarOps write
layer used by the assistant too, and reads that merge personal + shared events:
- domains/calendar/validators.ts — attendee/person/googleEvent/eventDoc, moved
  out of schema.ts to break the circular import (schema + write path share them)
- domains/calendar/tables.ts     — calendars/events/sharedEvents/sharedCalendars/
  recurringSeries, composed into schema.ts
- domains/calendar/model.ts       — selectedCalendarIds, EventView, shared-event
  reads
- domains/calendar/queries.ts     — read handlers (list/range/byId/recurrence/
  context/primary)
- domains/calendar/mutations.ts   — our-side writes (toggle, mirror, series cache)
- domains/calendar/service.ts     — the calendarOps ops (relocated last commit) +
  the write-action handlers appended
- calendar.ts (root)              — thin facade at the same paths/validators;
  re-exports EventView for the web app

googleSync's googleEventValidator import repointed to the new location.
Behavior-preserving: full suite green (calendar service 20 tests, booking itests
exercising upsertEvent/getPrimaryCalendarId), api.d.ts purely additive, apps/web
builds. Every api.calendar.* / internal.calendar.* path unchanged.
