+++ packages/backend/convex/schema.ts
@@ -2,109 +2,14 @@ import { defineSchema, defineTable } from "convex/server";
 import { v } from "convex/values";
 
 import { bookingTables } from "./domains/booking/tables";
+import { calendarTables } from "./domains/calendar/tables";
 import { marketingTables } from "./domains/marketing/tables";
 import { notificationTables } from "./domains/notifications/tables";
 import { peopleTables } from "./domains/people/tables";
 
-/** A guest on an event. Mirrors the subset of Google's attendee object we keep;
- * `responseStatus` is "needsAction" | "declined" | "tentative" | "accepted".
- * Shared by the `events` table and the mutation validators that write to it. */
-export const attendeeValidator = v.object({
-  email: v.string(),
-  displayName: v.optional(v.string()),
-  responseStatus: v.optional(v.string()),
-  organizer: v.optional(v.boolean()),
-  self: v.optional(v.boolean()),
-  optional: v.optional(v.boolean()),
-});
-
-/** A single person Google names on an event — its organizer or its creator.
- * Unlike an attendee this is not a guest list entry, so it carries no RSVP.
- * `self` is set by Google when the person is the calendar this copy lives on,
- * which makes it authoritative: never infer ownership by matching emails. */
-export const personValidator = v.object({
-  email: v.optional(v.string()),
-  displayName: v.optional(v.string()),
-  self: v.optional(v.boolean()),
-});
-
-/** One event exactly as Google gave it to us, after mapping. Shared by the
- * `events` table and by every action→mutation boundary that pushes a
- * MappedEvent across, so the shape is declared once and cannot drift. */
-export const googleEventValidator = v.object({
-  googleEventId: v.string(),
-  calendarId: v.string(),
-  summary: v.optional(v.string()),
-  description: v.optional(v.string()),
-  location: v.optional(v.string()),
-  startMs: v.number(),
-  endMs: v.number(),
-  allDay: v.boolean(),
-  status: v.string(),
-  htmlLink: v.optional(v.string()),
-  // Google's per-event color override ("1".."11"); absent means the event
-  // inherits its calendar's color.
-  colorId: v.optional(v.string()),
-  // Google's `visibility`: "default" | "public" | "private" | "confidential".
-  visibility: v.optional(v.string()),
-  // Google's `transparency`: "opaque" (busy) | "transparent" (free).
-  transparency: v.optional(v.string()),
-  // Guests invited to the event, refreshed by every sync. See attendeeValidator.
-  attendees: v.optional(v.array(attendeeValidator)),
-  // Google set this when it withheld part of the guest list, so a short
-  // `attendees` array is not proof that the list is short.
-  attendeesOmitted: v.optional(v.boolean()),
-  googleUpdatedMs: v.number(),
-
-  // --- Who controls this event -------------------------------------------
-  // `organizer.self` is the ownership test; `creator` differs when someone
-  // scheduled on the organizer's behalf, and is only worth showing then.
-  organizer: v.optional(personValidator),
-  creator: v.optional(personValidator),
-  // The three guest permissions. Each is tri-state on purpose: Google omits
-  // them at their default, and the defaults disagree — absent guestsCanModify
-  // means false, absent guestsCanInviteOthers/SeeOtherGuests mean true. Read
-  // them only through lib/permissions.ts, which encodes that.
-  guestsCanModify: v.optional(v.boolean()),
-  guestsCanInviteOthers: v.optional(v.boolean()),
-  guestsCanSeeOtherGuests: v.optional(v.boolean()),
-  // Google forbids structural changes to a locked event, whoever you are.
-  locked: v.optional(v.boolean()),
-  // "default" | "birthday" | "outOfOffice" | "focusTime" | "workingLocation" |
-  // "fromGmail". Several of these are generated by Google and reject edits even
-  // on a calendar you own.
-  eventType: v.optional(v.string()),
-  // Set on an expanded instance of a recurring series (we sync with
-  // singleEvents=true, so we only ever hold instances, never the master).
-  recurringEventId: v.optional(v.string()),
-  // The Google Meet URL, when the event has one.
-  hangoutLink: v.optional(v.string()),
-  // Provider-neutral conference metadata. Google Meet also remains available
-  // through hangoutLink for existing create/edit behavior.
-  conferenceUrl: v.optional(v.string()),
-  conferenceName: v.optional(v.string()),
-  conferenceType: v.optional(v.string()),
-});
-
-/** The stored row: everything Google told us, plus the local user it belongs
- * to. `userId` is ours alone — it never round-trips to Google. */
-export const eventDocValidator = googleEventValidator.extend({
-  userId: v.string(),
-  // Monotonic per-calendar full-resync marker. A full resync stamps every
-  // re-fetched row with a fresh generation, then deletes the rows still carrying
-  // an older one — so the previous snapshot stays live for booking conflict
-  // detection until the new one is fully written (see syncOneCalendar). Left
-  // absent on incrementally-written rows; the next full resync re-stamps any that
-  // still exist in Google, so absence never causes a wrongful sweep.
-  syncGeneration: v.optional(v.number()),
-  // Provider-neutral fields, dual-written alongside the Google-named columns
-  // above during the connection-model migration and read with legacy fallback
-  // until cutover. `providerEventId` mirrors `googleEventId`, `providerUpdatedMs`
-  // mirrors `googleUpdatedMs`. Optional until backfilled. See Stage 5.
-  connectionId: v.optional(v.id("calendarConnections")),
-  providerEventId: v.optional(v.string()),
-  providerUpdatedMs: v.optional(v.number()),
-});
+// The event validators (attendee/person/googleEvent/eventDoc) live in
+// domains/calendar/validators.ts so the schema and the calendar write path can
+// share them without a circular import back through this file.
 
 /** One piece of an assistant turn, in the order it happened.
  *
@@ -184,114 +89,9 @@ export default defineSchema({
     .index("by_user", ["userId"])
     .index("by_nextSyncDueAt", ["nextSyncDueAt"]),
 
-  // One row per Google calendar in the user's CalendarList. Holds display
-  // metadata, the user's visibility choice, and the per-calendar sync token.
-  calendars: defineTable({
-    userId: v.string(),
-    googleCalendarId: v.string(),
-    summary: v.optional(v.string()),
-    summaryOverride: v.optional(v.string()),
-    backgroundColor: v.optional(v.string()),
-    foregroundColor: v.optional(v.string()),
-    primary: v.optional(v.boolean()),
-    accessRole: v.optional(v.string()),
-    timeZone: v.optional(v.string()),
-    // Google's own Calendar UI visibility. This only seeds the local choice
-    // when a calendar is first discovered; later local toggles are preserved.
-    googleSelected: v.optional(v.boolean()),
-    selected: v.boolean(),
-    syncToken: v.optional(v.string()),
-    lastSyncAt: v.optional(v.number()),
-    // Current full-resync generation for this calendar's `events` rows. Bumped
-    // at the start of each full resync; the run stamps re-fetched rows with it
-    // and sweeps rows carrying an older value. See syncOneCalendar.
-    syncGeneration: v.optional(v.number()),
-    // Provider-neutral fields (dual-written until cutover). `providerCalendarId`
-    // mirrors `googleCalendarId`; `syncCursor` is the opaque per-calendar cursor
-    // that mirrors `syncToken` (a Google sync token today, a Graph delta link
-    // later). Optional until backfilled.
-    connectionId: v.optional(v.id("calendarConnections")),
-    providerCalendarId: v.optional(v.string()),
-    syncCursor: v.optional(v.string()),
-  })
-    .index("by_user", ["userId"])
-    .index("by_user_and_googleCalendarId", ["userId", "googleCalendarId"])
-    // Staged neutral-id lookup; read by nothing until cutover.
-    .index("by_connection_and_providerCalendarId", [
-      "connectionId",
-      "providerCalendarId",
-    ]),
-
-  // One row per synced Google Calendar event. See eventDocValidator above.
-  events: defineTable(eventDocValidator)
-    .index("by_user_and_start", ["userId", "startMs"])
-    // A dedicated (userId, calendarId) index would be redundant: every query
-    // that only constrains those two reuses by_user_and_calendar_and_end by
-    // prefix. Dropping it cuts a full index copy of the largest table.
-    .index("by_user_and_calendar_and_end", ["userId", "calendarId", "endMs"])
-    .index("by_user_and_calendar_and_googleEventId", [
-      "userId",
-      "calendarId",
-      "googleEventId",
-    ])
-    // Provider-neutral event-id lookup, the successor to the Google-id index
-    // above. Staged: built now but read by nothing until cutover, after which the
-    // Google-id index is retired. Costs one extra index copy on the largest table
-    // for the duration of the migration — an accepted, temporary cost.
-    .index("by_connection_and_providerEventId", [
-      "connectionId",
-      "providerEventId",
-    ]),
-
-  // One physical copy of a Google *public* calendar's events (holidays,
-  // birthdays — IDs ending "@group.v.calendar.google.com"), shared across every
-  // user who selects that calendar. Google generates this content identically
-  // for all viewers, so storing it once instead of per-user removes duplication
-  // that otherwise grows linearly with the user count. No `userId`: the row
-  // belongs to the calendar, not a person. See isSharedPublicCalendar in
-  // lib/calendars.ts and readVisibleEvents in calendar.ts.
-  sharedEvents: defineTable(googleEventValidator)
-    .index("by_calendar_and_start", ["calendarId", "startMs"])
-    .index("by_calendar_and_end", ["calendarId", "endMs"])
-    .index("by_calendar_and_googleEventId", ["calendarId", "googleEventId"]),
-
-  // One row per shared public calendar: its user-independent sync token plus a
-  // lease so exactly one user's sync refreshes it at a time. The per-user
-  // `calendars` row still holds display metadata and the `selected` toggle; only
-  // the events and their sync cursor are shared.
-  sharedCalendars: defineTable({
-    googleCalendarId: v.string(),
-    syncToken: v.optional(v.string()),
-    lastSyncAt: v.optional(v.number()),
-    // Held while a sync runs; a second user finding a live lease skips its run.
-    syncLeaseExpiresAt: v.optional(v.number()),
-  }).index("by_googleCalendarId", ["googleCalendarId"]),
-
-  // One row per recurring master. Expanded event instances share this rule;
-  // keeping it separately avoids duplicating it across every occurrence.
-  recurringSeries: defineTable({
-    userId: v.string(),
-    calendarId: v.string(),
-    googleEventId: v.string(),
-    recurrence: v.array(v.string()),
-    // The instance update time that this rule was fetched against. A newer
-    // synced instance invalidates the cache and triggers one master refresh.
-    sourceUpdatedMs: v.number(),
-    // Provider-neutral fields (dual-written until cutover). `providerEventId`
-    // mirrors `googleEventId`. Optional until backfilled.
-    connectionId: v.optional(v.id("calendarConnections")),
-    providerEventId: v.optional(v.string()),
-  })
-    .index("by_user_and_calendar_and_googleEventId", [
-      "userId",
-      "calendarId",
-      "googleEventId",
-    ])
-    // Staged neutral-id lookup; read by nothing until cutover.
-    .index("by_connection_and_providerEventId", [
-      "connectionId",
-      "providerEventId",
-    ]),
+  // Calendar domain tables — calendars / events / sharedEvents / sharedCalendars /
+  // recurringSeries (see domains/calendar/tables.ts).
+  ...calendarTables,
 
   // Booking domain tables — pages, date overrides, requests (see
   // domains/booking/tables.ts).
